v0.6.2: fix error
